working on assign subjet to class screen

+++ lib/models/teacher_model.dart
@@ -100,7 +100,7 @@ class Statistics {
   ByDepartment? byDepartment;
   ByStatus? byStatus;
   ByEmploymentType? byEmploymentType;
-  int? averageSalary;
+  double? averageSalary;
   ClassTeachers? classTeachers;
 
   Statistics({
